alpine: New package (v2)
Add Alpine v. 2.10 console mail client as new package to OpenWrt.

Alpine stands for Alternatively Licensed Program for Internet News and
Email and is suitable for both the inexperienced email user as well as
for the most demanding power user.

This is a v2 of my previous patch fixing build issues in certain
flavours of Linux. Thanks to Szymon Zygmunt for reporting.

Signed-off-by: Antti Seppälä <a.seppala@gmail.com>