telepathy-python: fix build error by updating to latest upstream version
The telepathy-python 0.15.19 tarball does not contain all files which
are referenced in the file ./spec/all.xml of the same tarball.

xsltproc < v1.1.27-rc1 shows an error about missing files, but goes ahead.
Starting with xsltproc 1.1.27-rc1 [1] an error code is returned, which
breaks the build.

[1] https://git.gnome.org/browse/libxslt/commit/xsltproc?id=e669a8c7cefec93ee1bfaf59721aadc42e226d6e

The current Ubuntu version already ships xsltproc 1.1.28.

Signed-off-by: Mathias Kresin <openwrt@kresin.me>